refactor: shared 폴더 제거
