+++ src/pages/OrderCheckoutPage/OrderCheckoutPage.js
@@ -1,19 +1,13 @@
 import { useHistory, useLocation } from 'react-router';
 import { useSelector } from 'react-redux';
 
-import {
-  Container,
-  CheckoutListContainer,
-  CheckoutList,
-  CheckoutListTitle,
-  PaymentInfoBoxContainer,
-} from './OrderCheckoutPage.styles';
+import { Container, CheckoutList, CheckoutListTitle, PaymentInfoBoxContainer } from './OrderCheckoutPage.styles';
 import { CONFIRM_MESSAGE, ROUTE, SCHEMA } from '../../constants';
 import { useServerAPI } from '../../hooks';
-import { numberWithCommas } from '../../shared/utils';
+import { numberWithCommas } from '../../utils';
 import { PaymentInfoBox, RowProductItem } from '../../components';
-import ScreenContainer from '../../shared/styles/ScreenContainer';
-import PageHeader from '../../shared/styles/PageHeader';
+import ScreenContainer from '../../styles/ScreenContainer';
+import PageHeader from '../../styles/PageHeader';
 
 const OrderCheckoutPage = () => {
   const history = useHistory();
@@ -59,7 +53,7 @@ const OrderCheckoutPage = () => {
       <PageHeader>주문/결제</PageHeader>
 
       <Container>
-        <CheckoutListContainer>
+        <>
           <CheckoutListTitle>{`주문 상품 ( ${checkedItemList.length || 0}건 )`}</CheckoutListTitle>
 
           <CheckoutList>
@@ -69,8 +63,8 @@ const OrderCheckoutPage = () => {
               return <RowProductItem key={id} imgSrc={img} name={name} amount={`수량: ${amount} 개`} />;
             })}
           </CheckoutList>
-        </CheckoutListContainer>
-
+        </>
+        CheckoutListContainer
         <PaymentInfoBoxContainer>
           <PaymentInfoBox
             title="결제금액"
