+++ src/styles/ScreenContainer.js
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import { ROUTE } from '../../constants';
+import { ROUTE } from '../constants';
 
 const ScreenContainer = styled.div`
   background-color: ${props => (props.route === ROUTE.ORDER_LIST ? 'rgba(229, 229, 229, 0.9)' : 'transparent')};
